reverted to version 1.14 due to compilation failure. PLEASE CLEAN YOUR TREE BEFORE RECOMPILING WHEN U CHANGE METHOD SIGNATURES
git-svn-id: file:///svn/trunk@2768 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/action/FundingAdded.java
@@ -70,9 +70,9 @@ public class FundingAdded extends Action {
 			}
 		}
 
-		long totComm = eaForm.getTotalCommitments();
-		long totDisb = eaForm.getTotalDisbursements();
-		long totExp = eaForm.getTotalExpenditures();
+		double totComm = eaForm.getTotalCommitments();
+		double totDisb = eaForm.getTotalDisbursements();
+		double totExp = eaForm.getTotalExpenditures();
 
 		Funding newFund = new Funding();
